change decorator to function
